requirements.txt
agency-swarm[fastapi]>=1.8.0
python-dotenv==1.1.1
